services/core/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.5.0
    # via django
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery==5.2.6
    # via
    #   -r requirements.in
    #   django-celery-results
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==36.0.2
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
deprecated==1.2.13
    # via redis
django==4.0.4
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-debug-toolbar
    #   django-filter
    #   django-model-utils
    #   django-modeltranslation
    #   djangorestframework
django-allauth==0.50.0
    # via -r requirements.in
django-celery-results==2.3.1
    # via -r requirements.in
django-debug-toolbar==3.2.4
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-model-utils==4.2.0
    # via -r requirements.in
django-modeltranslation==0.17.5
    # via -r requirements.in
django-simple-history==3.0.0
    # via -r requirements.in
django-webpack-loader==1.5.0
    # via -r requirements.in
django-widget-tweaks==1.4.12
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
kombu==5.2.4
    # via celery
markdown==3.3.6
    # via -r requirements.in
oauthlib==3.2.0
    # via requests-oauthlib
packaging==21.3
    # via redis
pillow==9.1.0
    # via -r requirements.in
prompt-toolkit==3.0.29
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.3.0
    # via django-allauth
pyparsing==3.0.8
    # via packaging
python3-openid==3.2.0
    # via django-allauth
pytz==2022.1
    # via
    #   celery
    #   djangorestframework
redis==4.2.2
    # via -r requirements.in
requests==2.27.1
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
six==1.16.0
    # via
    #   click-repl
    #   django-modeltranslation
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
urllib3==1.26.9
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
wrapt==1.14.0
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
